ai_engine/requirements.txt
flask
flask-cors
PyPDF2
python-dotenv
google-generativeai
